@@ -54,20 +54,20 @@
   "homepage": "https://jss.sitecore.net",
   "license": "Apache-2.0",
   "dependencies": {
-    "@angular/animations": "~9.1.12",
-    "@angular/common": "~9.1.12",
-    "@angular/compiler": "~9.1.12",
-    "@angular/core": "~9.1.12",
-    "@angular/forms": "~9.1.12",
-    "@angular/platform-browser": "~9.1.12",
-    "@angular/platform-browser-dynamic": "~9.1.12",
-    "@angular/platform-server": "~9.1.12",
-    "@angular/router": "~9.1.12",
+    "@angular/animations": "~10.0.11",
+    "@angular/common": "~10.0.11",
+    "@angular/compiler": "~10.0.11",
+    "@angular/core": "~10.0.11",
+    "@angular/forms": "~10.0.11",
+    "@angular/platform-browser": "~10.0.11",
+    "@angular/platform-browser-dynamic": "~10.0.11",
+    "@angular/platform-server": "~10.0.11",
+    "@angular/router": "~10.0.11",
     "@ngx-meta/core": "~9.0.0",
     "@ngx-translate/core": "~11.0.1",
     "@ngx-translate/http-loader": "~4.0.0",
     "@sitecore-jss/sitecore-jss": "^14.0.1",
-    "@sitecore-jss/sitecore-jss-angular": "file:../../packages/sitecore-jss-angular",
+    "@sitecore-jss/sitecore-jss-angular": "^14.0.1",
     "@sitecore-jss/sitecore-jss-tracking": "^14.0.1",
     "apollo-angular": "~1.10.0",
     "apollo-angular-link-http-batch": "~1.9.0",
@@ -81,14 +81,14 @@
     "graphql-tag": "~2.10.0",
     "lodash": "^4.17.15",
     "rxjs": "~6.5.3",
-    "tslib": "^1.10.0",
+    "tslib": "^2.0.1",
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.901.12",
-    "@angular/cli": "~9.1.12",
-    "@angular/compiler-cli": "~9.1.12",
-    "@angular/language-service": "~9.1.12",
+    "@angular-devkit/build-angular": "^0.1000.6",
+    "@angular/cli": "~10.0.6",
+    "@angular/compiler-cli": "~10.0.11",
+    "@angular/language-service": "~10.0.11",
     "@sitecore-jss/sitecore-jss-angular-schematics": "^14.0.1",
     "@sitecore-jss/sitecore-jss-cli": "^14.0.1",
     "@sitecore-jss/sitecore-jss-dev-tools": "^14.0.1",
@@ -117,8 +117,8 @@
     "protractor": "~5.4.2",
     "ts-loader": "~6.2.0",
     "ts-node": "~8.4.1",
-    "tslint": "~5.20.0",
-    "typescript": "~3.8.3",
+    "tslint": "~6.1.3",
+    "typescript": "~3.9.7",
     "webpack-cli": "~3.3.9",
     "yaml-lint": "^1.2.4"
   }
